Updating sim tests, mjcf exporter for no collisions, and removing legacy exporters
